mydata/llm/llama/requirements.txt
fastapi
uvicorn
llama-cpp-python
torch
numpy
sentencepiece
transformers
huggingface_hub

